Fix uninitialized value warning in RT::Principal::HasRight
HasRight was storing the cananicalized value of a passed-in
Right argument back in the argument hash, so the debug log output
on an invalid right had an uninitialized value rather
than the invalid right argument.

Store the cananicalized right in a new variable to allow the debug
message to report on the invalid right passed in.
Also, avoid putting a modified argument back in the arg hash
which can be misleading.