Stop looking for things in the GnuPG configuration
This restores the behavior that if mail comes in which is "invalid" in
some way that we want to reject, we'll stop gpg/smime processing the
mail after notifying external users of the failure (basically skip the
parts that update X-RT-Incoming-(Signature|Encryption) to record
success).  I'm not sure if that's actually the optimal behavior, but
it's the back-compat behavior.