Generalize _CustomFieldJoin somewhat
This replaces explicit RT::Ticket ObjectTypes and LookupTypes with their
more general forms.  However, if a CF name (and not object) is passed
in, this only attempts to search global CFs of the appropriate object
type.  The logic to limit non-global CFs (for instance, joining
transactions to tickets to queues) does not exist yet.