Switch the general LimitCustomField to the new implementation
Use the new _LimitCustomField implementation for RT::SearchBuilder's
LimitCustomField.  The former is not simply renamed to the latter
because RT::Tickets contains a LimitCustomField which acts on
"restrictions", and must continue to do so.  As such, _LimitCustomField
(and its recursive calls) must continue to occupy some other name.

This removes the (dubious and unused) ability to call LimitCustomField
with no CUSTOMFIELD provided, which searches within all custom fields.
It adds all of the complex date and IP parsing capabilities that Tickets
have enjoyed.