Detect and warn of configurations with too small max_allowed_packet
MySQL's max_allowed_packet controls the largest ata that can be inserted
into the database; surpassing this causes inserts to fail.  This setting
defaults to 1M -- as RT stores attachments in the database, this is far
too small for most deployments.

Detect and warn of this situation, both when installing and at run-time.